requirements.txt
anyio==3.7.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==41.0.3
dnspython==2.3.0
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.99.1
h11==0.14.0
idna==3.4
motor==3.2.0
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.10
pymongo==4.4.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
